+++ src/main/kotlin/quickfix/dao/Repository.kt
@@ -6,7 +6,7 @@ import quickfix.models.Id
 
 abstract class Repository<T : Id> {
   protected val elements: MutableSet<T> = mutableSetOf()
-  private var currentId : Int = 0
+  private var currentId : Long = 0
 
   fun create(element: T) {
     throwErrorIfIdIsAssigned(element)
@@ -29,11 +29,11 @@ abstract class Repository<T : Id> {
   private fun addElement(element: T) {
     elements.add(element)
   }
-  fun getById(id : Int) : T? {
+  fun getById(id : Long) : T? {
     throwErrorIfIdDoesNotExist(id)
     return elements.find { it.id == id }
   }
-  private fun throwErrorIfIdDoesNotExist(id : Int){
+  private fun throwErrorIfIdDoesNotExist(id : Long){
     if(elements.all{it.id != id})
       throw BusinessException("There is no element associated with the id:: $id")
   }
